Move pid/tid handling to OrbitBase.
I need these outside of LinuxTracing.

Bug: N/A
Test: Unit tests.
