+++ src/OrbitBase/CMakeLists.txt
@@ -21,16 +21,17 @@ target_include_directories(OrbitBase PRIVATE
 target_sources(OrbitBase PRIVATE
         include/OrbitBase/Action.h
         include/OrbitBase/ExecutablePath.h
-        include/OrbitBase/ReadFileToString.h
         include/OrbitBase/Logging.h
         include/OrbitBase/MakeUniqueForOverwrite.h
+        include/OrbitBase/ProcessId.h
         include/OrbitBase/Profiling.h
-        include/OrbitBase/UniqueResource.h
+        include/OrbitBase/ReadFileToString.h
+        include/OrbitBase/SafeStrerror.h
         include/OrbitBase/ThreadConstants.h
         include/OrbitBase/ThreadPool.h
         include/OrbitBase/ThreadUtils.h
         include/OrbitBase/Tracing.h
-        include/OrbitBase/SafeStrerror.h)
+        include/OrbitBase/UniqueResource.h)
 
 target_sources(OrbitBase PRIVATE
         ExecutablePath.cpp
@@ -47,6 +48,7 @@ target_sources(OrbitBase PRIVATE
 else()
 target_sources(OrbitBase PRIVATE
         ExecutablePathLinux.cpp
+        ProcessIdLinux.cpp
         ThreadUtilsLinux.cpp)
 endif()
 
@@ -72,7 +74,8 @@ target_sources(OrbitBaseTests PRIVATE
 
 if (NOT WIN32)
 target_sources(OrbitBaseTests PRIVATE
-        ExecutablePathLinuxTest.cpp)
+        ExecutablePathLinuxTest.cpp
+        ProcessIdLinuxTest.cpp)
 endif()
 
 # Threadpool test contains some sleeps we couldn't work around - disable them on the CI
